refactor(work tracker service): relocate to a more generic location
WorkTrackerService was previouly owned by the picklist, but relocating to /shared since app switcher
component uses the service as well
