+++ projects/cashmere/src/lib/app-switcher/app-switcher.module.ts
@@ -6,7 +6,7 @@ import {AppSwitcherComponent} from './app-switcher.component';
 import {PopModule} from '../pop/popover.module';
 import {ProgressIndicatorsModule} from '../progress-indicators/progress-indicators.module';
 import {AppSwitcherService} from './app-switcher.service';
-import { WorkTrackerService } from '../picklist/services/work-tracker.service';
+import { WorkTrackerService } from '../shared/work-tracker.service';
 import {PipesModule} from '../pipes/pipes.module';
 import {IAppSwitcherConfig, APP_SWITCHER_CONFIG, APP_SWITCHER_SERVICE} from './app-switcher-interfaces';
 import {AppSwitcherLinksComponent} from './app-switcher-application-link/app-switcher-links.component';
